+++ build.sh
@@ -8,7 +8,7 @@ echo 'Run testing'
 npm test
 npm run build
 
-if [ "${TRAVIS_NODE_VERSION}" != "10.14.2" ]; then
+if [ "${TRAVIS_NODE_VERSION}" != "10.15.3" ]; then
     exit 0
 fi