+++ src/components/common/tooltips/Tooltip.js
@@ -17,12 +17,15 @@
 
 /* global requestAnimationFrame */
 
-import React, { PropTypes, PureComponent } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import _ from 'lodash';
+import moment from 'moment-timezone';
 
+import { formatLocalizedFromUTC, getHourMinuteFormat } from '../../../utils/datetime';
 import styles from './Tooltip.css';
 
-class Tooltip extends PureComponent {
+class Tooltip extends React.PureComponent {
   constructor(props) {
     super(props);
     this.state = { offset: { top: 0, left: 0 } };
@@ -148,28 +151,72 @@ class Tooltip extends PureComponent {
     );
   }
 
-  renderTitle(title) {
-    const { tail, content } = this.props;
+  renderTitle() {
+    const { title, dateTitle, tail, content } = this.props;
     let renderedTitle = null;
+    let tailNode = null;
+    if (tail && content === null) {
+      tailNode = this.renderTail(styles.tooltipTitleBg);
+    }
     if (title) {
       renderedTitle = (
         <div className={styles.title}>
           <span>{title}</span>
-          {tail && !content && this.renderTail(styles.tooltipTitleBg)}
+          {tailNode}
+        </div>
+      );
+    } else if (dateTitle) {
+      let titleNode = null;
+      if (dateTitle.source === 'Diabeloop') {
+        // For diabeloop device, use the timezone of the object
+        const { timezoneName, timezoneOffset } = dateTitle.timePrefs;
+        const mNormalTime = moment.tz(dateTitle.normalTime, dateTitle.timezone);
+        // const timelineTimezone = getTimezoneFromTimePrefs(dateTitle.timePrefs);
+        let displayOffset = null;
+        if (timezoneName !== dateTitle.timezone) {
+          // Not the same timezone than the timePrefs, so make it visible
+          displayOffset = ` UTC${mNormalTime.format('Z')}`;
+        } else if (timezoneOffset !== mNormalTime.utcOffset()) {
+          // Not the same offset than the current one
+          displayOffset = ` UTC${mNormalTime.format('Z z')}`;
+        }
+        titleNode = (
+          <div className={styles.title}>
+            {mNormalTime.format(getHourMinuteFormat())}
+            {displayOffset}
+          </div>
+        );
+      } else {
+        // eslint-disable-next-line max-len
+        const time = formatLocalizedFromUTC(dateTitle.normalTime, dateTitle.timePrefs, getHourMinuteFormat());
+        titleNode = (
+          <div className={styles.title}>
+            {time}
+          </div>
+        );
+      }
+      renderedTitle = (
+        <div className={styles.title}>
+          <span>{titleNode}</span>
+          {tailNode}
         </div>
       );
     }
     return renderedTitle;
   }
 
-  renderContent(content) {
+  renderContent() {
+    const { tail, content } = this.props;
     let renderedContent = null;
-    const { tail } = this.props;
     if (content) {
+      let tailNode = null;
+      if (tail) {
+        tailNode = this.renderTail();
+      }
       renderedContent = (
         <div className={styles.content}>
           <span>{content}</span>
-          {tail && this.renderTail()}
+          {tailNode}
         </div>
       );
     }
@@ -177,7 +224,7 @@ class Tooltip extends PureComponent {
   }
 
   render() {
-    const { title, content, position, backgroundColor, borderColor, borderWidth } = this.props;
+    const { position, backgroundColor, borderColor, borderWidth } = this.props;
     const { offset } = this.state;
     const top = position.top + offset.top;
     const left = position.left + offset.left;
@@ -188,8 +235,8 @@ class Tooltip extends PureComponent {
         style={{ top, left, backgroundColor, borderColor, borderWidth: `${borderWidth}px` }}
         ref={this.setElementRef}
       >
-        {title && this.renderTitle(title)}
-        {content && this.renderContent(content)}
+        {this.renderTitle()}
+        {this.renderContent()}
       </div>
     );
   }
@@ -199,6 +246,16 @@ Tooltip.displayName = 'Tooltip';
 
 Tooltip.propTypes = {
   title: PropTypes.node,
+  dateTitle: PropTypes.shape({
+    normalTime: PropTypes.string.isRequired,
+    timezone: PropTypes.string.isRequired,
+    source: PropTypes.string.isRequired,
+    timePrefs: PropTypes.shape({
+      timezoneAware: PropTypes.bool.isRequired,
+      timezoneName: PropTypes.string.isRequired,
+      timezoneOffset: PropTypes.number.isRequired,
+    }).isRequired,
+  }),
   content: PropTypes.node,
   position: PropTypes.shape({
     top: PropTypes.number.isRequired,
@@ -209,7 +266,7 @@ Tooltip.propTypes = {
     left: PropTypes.number,
     horizontal: PropTypes.number,
   }).isRequired,
-  tail: PropTypes.bool.isRequired,
+  tail: PropTypes.bool,
   side: PropTypes.oneOf(['top', 'right', 'bottom', 'left']).isRequired,
   tailWidth: PropTypes.number.isRequired,
   tailHeight: PropTypes.number.isRequired,
@@ -220,6 +277,7 @@ Tooltip.propTypes = {
 };
 
 Tooltip.defaultProps = {
+  content: null,
   tail: true,
   side: 'left',
   tailWidth: 7,
@@ -227,6 +285,8 @@ Tooltip.defaultProps = {
   borderColor: 'black',
   borderWidth: 2,
   offset: { top: 0, left: 0 },
+  title: null,
+  dateTitle: null,
 };
 
 export default Tooltip;