+++ src/components/daily/parametertooltip/ParameterTooltip.js
@@ -17,15 +17,14 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
+import _ from 'lodash';
+import i18next from 'i18next';
 
 import { formatLocalizedFromUTC, getHourMinuteFormat } from '../../../utils/datetime';
 import { formatParameterValue } from '../../../utils/format';
 import Tooltip from '../../common/tooltips/Tooltip';
 import colors from '../../../styles/colors.css';
 import styles from './ParameterTooltip.css';
-import i18next from 'i18next';
-
-const t = i18next.t.bind(i18next);
 
 class ParameterTooltip extends React.Component {
   constructor(props) {
@@ -53,11 +52,11 @@ class ParameterTooltip extends React.Component {
 
     return [
       <span className={styles.date} key={`${parameter.id}-date`}>{displayHour}</span>,
-      <span className={styles.label} key={`${parameter.id}-name`}>{t(`params:::${parameter.name}`)}</span>,
+      <span className={styles.label} key={`${parameter.id}-name`}>{i18next.t(`params:::${parameter.name}`)}</span>,
       previousValue,
       prevToNext,
       <span className={valueClassName} key={`${parameter.id}-value`}>{value}</span>,
-      <span className={styles.units} key={`${parameter.id}-units`}>{t(`${parameter.units}`)}</span>,
+      <span className={styles.units} key={`${parameter.id}-units`}>{i18next.t(`${parameter.units}`)}</span>,
     ];
   }
 
@@ -75,21 +74,24 @@ class ParameterTooltip extends React.Component {
   }
 
   render() {
-    const title = this.props.title ? this.props.title : (
-      <div className={styles.title}>
-        {
-          formatLocalizedFromUTC(
-            this.props.parameter.normalTime,
-            this.props.timePrefs,
-            this.hourMinuteFormat)
-          }
-      </div>
-    );
-    const content = this.renderParameters(this.props.parameter.params);
+    const { parameter, timePrefs, title } = this.props;
+
+    let dateTitle = null;
+    if (title === null) {
+      dateTitle = {
+        source: _.get(parameter, 'source', 'tidepool'),
+        normalTime: parameter.normalTime,
+        timezone: _.get(parameter, 'timezone', 'UTC'),
+        timePrefs,
+      };
+    }
+
+    const content = this.renderParameters(parameter.params);
     return (
       <Tooltip
         {...this.props}
         title={title}
+        dateTitle={dateTitle}
         content={content}
       />
     );
@@ -138,6 +140,7 @@ ParameterTooltip.defaultProps = {
   tailColor: colors.deviceEvent,
   borderColor: colors.deviceEvent,
   borderWidth: 2,
+  title: null,
 };
 
 export default ParameterTooltip;