+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 This node package is a dependency of Blip and was forked from tidepool/viz.
 
 ## Unreleased
+### Changed
+- PT-1205 Display data based on timezone objects
 ### Fixed
 - PT-1244 Parameter names are not aligned in Daily tooltip
 - PT-1245 Parameter values are not rounded in Daily tooltip and Device settings page